django_app/requirements.txt
django
kafka-python
